added authentication/authorization,JWT and more videos

+++ src/main/java/com/nano/videosite/controllers/PlaylistController.java
@@ -12,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -19,6 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.nano.videosite.models.Playlist;
 import com.nano.videosite.models.Video;
+import com.nano.videosite.services.JWTAuthenticationService;
 import com.nano.videosite.services.PlaylistService;
 
 @RestController
@@ -26,7 +28,8 @@ public class PlaylistController {
 	
 	@Autowired
 	PlaylistService playlistService;
-	
+	@Autowired
+	JWTAuthenticationService jwtService;
 //	@RequestMapping(method=RequestMethod.GET, value="/user/{id}/playlist", produces={MediaType.APPLICATION_JSON_VALUE})
 //	public ResponseEntity<Set<Map<Integer,Video>>> allPlaylists(@PathVariable("id") Long userId){
 //		Set<Map<Integer,Video>> playlists = playlistService.allPlaylists(userId);
@@ -40,9 +43,13 @@ public class PlaylistController {
 	}
 	
 	@RequestMapping(method=RequestMethod.POST, value="/playlist", produces={MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<Playlist> add(@RequestBody Playlist newPlaylist) {
-		Playlist playlist = playlistService.add(newPlaylist);
-		return ResponseEntity.status(HttpStatus.CREATED).body(playlist);
+	public ResponseEntity<Playlist> add(@RequestBody Playlist newPlaylist, @RequestHeader("Authorization") String authorization) {
+		if(jwtService.isAuthenticated(authorization)) {
+			Playlist playlist = playlistService.add(newPlaylist);
+			return ResponseEntity.status(HttpStatus.CREATED).body(playlist);
+		}else {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
 	}
 	
 	@RequestMapping(method=RequestMethod.GET, value="/playlist/list/{id}", produces={MediaType.APPLICATION_JSON_VALUE})
@@ -62,34 +69,52 @@ public class PlaylistController {
 	}
 	
 	@RequestMapping(method = RequestMethod.PUT, value="/playlist/{id}/edit/order-change", produces={MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<Map<Integer, Video>> editOrder(@RequestBody Map<Integer,Video> newPlaylist, @PathVariable("id") Long playlistId){
-		Map<Integer, Video> playlist = playlistService.editOrder(playlistId,newPlaylist);
-		//Or I could return 204, with no Response Body. For Testing, return a body.
-//		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-		return ResponseEntity.ok(playlist);
+	public ResponseEntity<Map<Integer, Video>> editOrder(@RequestBody Map<Integer,Video> newPlaylist, @PathVariable("id") Long playlistId, @RequestHeader("Authorization") String authorization){
+		if(jwtService.isAuthenticated(authorization)) {
+			Map<Integer, Video> playlist = playlistService.editOrder(playlistId,newPlaylist);
+			//Or I could return 204, with no Response Body. For Testing, return a body.
+//			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+			return ResponseEntity.ok(playlist);
+		}else {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
+		
 	}
 	
 	@RequestMapping(method = RequestMethod.PUT, value="/playlist/{id}/edit/title-change", produces={MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<Playlist> editTitle(@RequestBody Playlist newPlaylist,@PathVariable("id") Long playlistId){
-		Playlist playlist = playlistService.editTitle(playlistId, newPlaylist);
-		//Or I could return 204, with no Response Body. For Testing, return a body.
-//		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-		return ResponseEntity.ok(playlist);
+	public ResponseEntity<Playlist> editTitle(@RequestBody Playlist newPlaylist,@PathVariable("id") Long playlistId, @RequestHeader("Authorization") String authorization){
+		if(jwtService.isAuthenticated(authorization)) {
+			Playlist playlist = playlistService.editTitle(playlistId, newPlaylist);
+			//Or I could return 204, with no Response Body. For Testing, return a body.
+//			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+			return ResponseEntity.ok(playlist);
+		}else {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
+		
 	}
 	
 	@RequestMapping(method = RequestMethod.POST, value="/playlist/{id}/edit/add-video", produces={MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<Playlist> editAddVideo(@RequestBody List<Video> newVideo,@PathVariable("id") Long playlistId){
-		Playlist playlist = playlistService.editAddVideo(playlistId, newVideo);
+	public ResponseEntity<Playlist> editAddVideo(@RequestBody List<Video> newVideo,@PathVariable("id") Long playlistId, @RequestHeader("Authorization") String authorization){
+		if(jwtService.isAuthenticated(authorization)) {
+			Playlist playlist = playlistService.editAddVideo(playlistId, newVideo);
+			return ResponseEntity.ok(playlist);
+		}else {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
 		//Or I could return 204, with no Response Body. For Testing, return a body.
 //		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-		return ResponseEntity.ok(playlist);
 	}
 	
 	@RequestMapping(method = RequestMethod.DELETE, value="/playlist/{id}")
-	public ResponseEntity deletePlaylist(@PathVariable("id") Long playlistId){
+	public ResponseEntity deletePlaylist(@PathVariable("id") Long playlistId,@RequestHeader("Authorization") String authorization){
 //		System.out.println("deleted playlist");
-		playlistService.deletePlaylist(playlistId);
-		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+		if(jwtService.isAuthenticated(authorization)) {
+			playlistService.deletePlaylist(playlistId);
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+		}else {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
 	}
 }
 //Maybe I should use the below code for JWT authorization?
